info-agent-service/pyproject.toml
[project]
name = "info-agent-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "groqchat>=0.0.14",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-nomic>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "langgraph-checkpoint-postgres>=3.0.5",
    "motor>=3.7.1",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "redis>=7.4.0",
    "rich>=14.3.3",
    "uvicorn>=0.42.0",
]
